+++ examples/http_client_usage/simple_example.py
@@ -15,11 +15,11 @@ import pyjsonrpc
 rpc_client = pyjsonrpc.HttpClient(url = "http://localhost:8080")
 
 # Example with *call*
-print rpc_client.call("hello", u"Gerold from Österreich")
+print rpc_client.call("add", 1, 2)
 
 # Example with direct *call*
-print rpc_client("hello", u"Gerold from Österreich")
+print rpc_client("add", 1, 2)
 
 # Example with *attribute* name
-print rpc_client.hello(u"Gerold from Österreich")
+print rpc_client.add(1, 2)